handling-pdf-files/pdf-signer/requirements.txt
PDFNetPython3==8.1.0
pyOpenSSL==20.0.1